책 소개 동적 라우팅 -> 취소 -> 캐러셀 - 상세저보 연결 구현 완료
