+++ myBookReport/public/common/styles/common.css
@@ -808,6 +808,11 @@ header .scrollTop::after {
   margin-bottom: 20px;
 }
 
+.bookpage .rollimg.active .imgbox {
+  display: block;
+  border: solid 3px #d1e0e5;
+}
+
 /* 미디어 쿼리 */
 @media screen and (max-width: 1024px) {
   .system > button {
@@ -977,5 +982,96 @@ header .scrollTop::after {
   .tapcontit,
   .tapcontit > strong {
     font-size: 15px;
+  }
+
+  .bookMain {
+    display: flex;
+    flex-direction: column;
+    margin-top: 117px;
+  }
+
+  .carousel-item {
+    display: block;
+  }
+
+  .rollimg {
+    padding: 0 10px;
+  }
+
+  .imgbox > img {
+    max-width: 100%;
+    height: 200px;
+  }
+
+  .rollimg {
+    display: flex;
+  }
+  .rollimg > strong {
+    display: block;
+    font-size: 18px;
+    color: #666;
+  }
+
+  .bookname {
+    font-size: 14px;
+  }
+
+  .textbox {
+    display: block;
+    font-size: 12px;
+    color: #ababab;
+  }
+
+  .carousel-item {
+    display: block;
+  }
+
+  .subpageslider {
+    display: block;
+    padding: 20px 20px 10px 20px;
+    background-color: #f5f5f5;
+    height: 300px;
+  }
+
+  .subpageslider .rollimg > strong {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+
+  .buttons {
+    position: relative;
+    z-index: 1;
+    pointer-events: auto;
+  }
+
+  .leftButton {
+    position: absolute;
+    top: -150px;
+    left: 50px;
+    pointer-events: auto;
+    font-size: 24px;
+  }
+
+  .buttons > .leftButton:hover {
+    cursor: pointer;
+    color: black !important;
+  }
+
+  .rightButton {
+    position: absolute;
+    top: -150px;
+    left: 1100px;
+    font-size: 24px;
+  }
+
+  .rightButton:hover {
+    cursor: pointer;
+    color: black !important;
+  }
+
+  .bookTabs li {
+    width: 150px;
+    height: 45px;
   }
 }
